refactor(patina): split style directive lexer into its own module
