+++ crates/vize_patina/src/context/sfc_directives/lexer.rs
@@ -1,6 +1,9 @@
 //! Lightweight comment lexers for SFC script and style block contents.
 
+mod style;
 mod token;
+
+pub(super) use style::StyleDirectiveLexer;
 
 use token::{
     ends_with_unescaped_backslash, identifier_allows_expression, identifier_end,
@@ -262,91 +265,7 @@ impl DirectiveLexer {
     }
 }
 
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum StyleContext {
-    Code,
-    SingleQuote,
-    DoubleQuote,
-    BlockComment,
-    LineComment,
-}
-
-pub(super) struct StyleDirectiveLexer {
-    context: StyleContext,
-    allow_line_comments: bool,
-}
-
-impl StyleDirectiveLexer {
-    pub(super) fn new(allow_line_comments: bool) -> Self {
-        Self {
-            context: StyleContext::Code,
-            allow_line_comments,
-        }
-    }
-
-    pub(super) fn scan_line(&mut self, line: &str) -> CommentMarkers {
-        let bytes = line.as_bytes();
-        let mut markers = CommentMarkers::default();
-        let mut index = 0;
-        while index < bytes.len() {
-            if matches!(
-                self.context,
-                StyleContext::BlockComment | StyleContext::LineComment
-            ) {
-                record_markers(bytes, index, &mut markers);
-            }
-            let current = bytes[index];
-            let next = bytes.get(index + 1).copied();
-            match self.context {
-                StyleContext::Code => match (current, next) {
-                    (b'/', Some(b'*')) => {
-                        self.context = StyleContext::BlockComment;
-                        index += 1;
-                    }
-                    (b'/', Some(b'/')) if self.allow_line_comments => {
-                        self.context = StyleContext::LineComment;
-                        index += 1;
-                    }
-                    (b'\'', _) => self.context = StyleContext::SingleQuote,
-                    (b'"', _) => self.context = StyleContext::DoubleQuote,
-                    _ => {}
-                },
-                StyleContext::SingleQuote => match (current, next) {
-                    (b'\\', Some(_)) => index += 1,
-                    (b'\'', _) => self.context = StyleContext::Code,
-                    _ => {}
-                },
-                StyleContext::DoubleQuote => match (current, next) {
-                    (b'\\', Some(_)) => index += 1,
-                    (b'"', _) => self.context = StyleContext::Code,
-                    _ => {}
-                },
-                StyleContext::BlockComment => {
-                    if (current, next) == (b'*', Some(b'/')) {
-                        self.context = StyleContext::Code;
-                        index += 1;
-                    }
-                }
-                StyleContext::LineComment => {}
-            }
-            index += 1;
-        }
-        if self.context == StyleContext::LineComment {
-            self.context = StyleContext::Code;
-        }
-        if !ends_with_unescaped_backslash(bytes)
-            && matches!(
-                self.context,
-                StyleContext::SingleQuote | StyleContext::DoubleQuote
-            )
-        {
-            self.context = StyleContext::Code;
-        }
-        markers
-    }
-}
-
-fn record_markers(bytes: &[u8], index: usize, markers: &mut CommentMarkers) {
+pub(super) fn record_markers(bytes: &[u8], index: usize, markers: &mut CommentMarkers) {
     if markers.eslint.is_none() && bytes[index..].starts_with(b"eslint-") {
         markers.eslint = Some(index);
     }
